renames package to contact
